requirements.txt
#
# This file is autogenerated by pip-compile
# Make changes in requirements.in, then run this to update:
#
#    pip-compile requirements.in
#
astroid==1.4.4            # via pylint
colorama==0.3.6           # via pylint
Django==1.9.2
factory-boy==2.6.1
fake-factory==0.5.3       # via factory-boy
lazy-object-proxy==1.2.1  # via astroid
py==1.4.31                # via pytest
pylint==1.5.4
pytest==2.8.7
pytz==2015.7
six==1.10.0               # via astroid, pylint
wrapt==1.10.6             # via astroid
